LPS-58479 Move and rename for consistency
